Create MyGame.java
This is a Guessing Game in which the user has to enter any number between 1 and 100, after which the user's input is compared to the guessed value. If the user's guess is correct, then the score will be updated, else the number of attempts will be decremented and a message containing the feedback will be produced.